1. 移除部分空白行末尾的空白字符 2. 修正部分命名拼写错误（仅限于本地变量 和 private 属性 ） 3. 基于 Java 5 的自动装箱、拆箱特性、增强for循环，简化部分代码 4. 改进 部分泛型声明和泛型转换问题，移除冗余的类型转换，减少编译器警告 5. 移除部分冗余的 continue 和 return 语句 6. 移除部分冗余的变量声明和初始化（仅限于方法中的本地变量） 7. 复用流程控制语句中的重复代码 8. 进一步复用并优化重载方法中的重复代码 9. 移除字符串中不必要的转义符 \ 10. 使用 String.equals() 替换 == 11. 不应该写 javadoc /**  */的地方，改为 block comment /*  */ 12. 使用 JDK 内置的 Collections.addAll() 和 Arrays.fill() 等方法替代循环代码 13. 在不改变代码逻辑的情况下，精简部分代码

+++ src/main/java/com/alibaba/fastjson/serializer/DoubleSerializer.java
@@ -48,7 +48,7 @@ public class DoubleSerializer implements ObjectSerializer {
             return;
         }
 
-        double doubleValue = ((Double) object).doubleValue();
+        double doubleValue = (Double) object;
 
         if (Double.isNaN(doubleValue) //
                 || Double.isInfinite(doubleValue)) {
